Add files via upload
Hello everyone the other day I decided to write a program to check the password strength. Earlier I wrote similar programs, on github they are not patamous that the programs did not completely turn out the way I wanted. The previous program not only knew how to check the password and give out a generated strong password based on it, but also save the password in the user database. Honestly, I did this project in a hurry. I still have it, and maybe I'll post it on the github, although on the other hand it's a shame to look at the code that I wrote there. Well, in general, it was a relative digression, let's move on to inspect this program. My program has a minus over the previous one, namely, it cannot save user passwords. While writing this program, I learned how to design more complex algorithms, for example, an algorithm for generating a strong password. This algorithm is 2 times better than the algorithm in the previous program, since it generates different password variations. Also, when writing the program, I tried to adhere to the SOLID rules, but by the time I had to break one rule, this is the rule of the only responsibility. I violated his patam that when I made the language interface of the program, I had to work directly with the main class in which the executable method is located. Further plans for the program are to pack this into an executable file so that the program can be run on any computer without installing jre